[APPC-4627] Rename payment_methods package to payment_methods_list
